requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.0
asyncpg==0.29.0
flask==3.0.0